% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s216-applets.R
\name{add_norm}
\alias{add_norm}
\title{Add normal curve to histogram}
\usage{
add_norm(
  h,
  data,
  freq = TRUE,
  na.rm = FALSE,
  lty = 1,
  lwd = 2,
  col = "purple",
  ...
)
}
\arguments{
\item{h}{Histogram object generated by \code{\link[graphics:hist]{hist}}.}

\item{data}{Vector of values used to generate the histogram.}

\item{freq}{Logical value indicating whether y-axis of
histogram is frequency. Defaults to TRUE.}

\item{na.rm}{Logical value indicating whether NA values should
be stripped before the computation proceeds. Defaults to TRUE.}

\item{...}{Other \code{\link[graphics:par]{graphical parameters}}.}

\item{add_normal}{Logical value indicating whether to superimpose a normal
curve on the histogram. Defaults to FALSE.}
}
\description{
This function will add a normal curve to a frequency or density histogram
using the raw data and histogram object as arguments.
}
\examples{
x <- rnorm(30, 5, 2)
hi <- hist(x)
add_norm(h = hi, data = x)
}
