% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ch5.R
\docType{data}
\name{BB}
\alias{BB}
\title{Blood Alcohol Contents}
\format{
A data frame with 16 rows and 2 variables:
\describe{
 \item{Beers}{number of cans of beer consumed}
 \item{BAC}{blood alcohol content, in grams per decaliter}
}
}
\source{
In progress
}
\usage{
BB
}
\description{
A group of n=16 student volunteers at The Ohio State University drank a
randomly assigned number of cans of beer. Thirty minutes later, a police
officer measured their BAC.
}
\keyword{datasets}
