% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/interaction-plot.R
\name{intplot}
\alias{intplot}
\title{Interaction Plots with Error Bars}
\usage{
intplot(formula = NULL, data = NULL, type = "b", x.cont = FALSE,
  legend = TRUE, trace.label = NULL, leg.lab = NULL, fixed = FALSE,
  x.leg = NULL, y.leg = NULL, cex.leg = 1, ncol = 1, pch = c(16, 21,
  15, 22, 17, 24, c(3:14)), fun = function(x) mean(x, na.rm = TRUE),
  ci.fun = function(x) c(fun(x) - se(x), fun(x) + se(x)), err.width = if
  (length(levels(as.factor(x.factor))) > 10) 0 else 0.1, err.col = col,
  err.lty = 1, xlim = NULL, ylim = NULL, cex = NULL, lwd = NULL,
  col = "black", cex.axis = 1, xaxt = "s", main = NULL, cld = F,
  cldshift = 0.1, cldcol = "white", ...)
}
\arguments{
\item{formula}{an object of class "\link{formula}" (or one that can be
coerced to that class); a symbolic description of the model to be plotted.
The details of the model specification are given under "Details".}

\item{data}{an optional data frame containing the variables in the model.  If
not found in \code{data}, the variables are taken from the global
environment (I'm assuming).}

\item{type}{character.  One of "l", "p", "b".  Defaults to "b".  Does
something}

\item{x.cont}{boolean.  Defaults to FALSE.}

\item{legend}{boolean.  Defaults to TRUE.}

\item{trace.label}{}

\item{leg.lab}{}

\item{fixed}{}

\item{x.leg}{}

\item{y.leg}{}

\item{cex.leg}{}

\item{ncol}{}

\item{pch}{}

\item{fun}{}

\item{ci.fun}{}

\item{err.width}{}

\item{err.col}{}

\item{err.lty}{}

\item{xlim}{}

\item{ylim}{}

\item{cex}{}

\item{lwd}{}

\item{col}{}

\item{cex.axis}{}

\item{xaxt}{}

\item{main}{}

\item{cld}{}

\item{cldshift}{}

\item{cldcol}{}

\item{...}{optional arguments to be passed to plot.}
}
\description{
This function allows you to generate interaction plots for a Two-Way ANOVA
with error bars.
}

