% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ch5.R
\docType{data}
\name{mtfires}
\alias{mtfires}
\title{Montana Fires}
\format{
A data frame with 23 rows and 3 columns:
\describe{
\item{Year}{year of data}
\item{Temperature}{average daily high temperature, in degrees Fahrenheit}
\item{Hectares}{total area burned by wildfires.  One hectare is equal to
10,000 square meters.}
}
}
\source{
Gude, P.H., Cookson, A.J., Greenwood, M.C., and Haggerty, M. (2009)
Homes in Wildfire-Prone Areas:An Empirical Analysis of Wildfire Suppression
Costs and Climate Change.
}
\usage{
mtfires
}
\description{
This data frame contains information about yearly wildfires in Montana from
1985 to 2007.  Each year constitutes one observation in these data so there
are 23 rows (one for each year).
}
\keyword{datasets}
