% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s216-applets.R
\name{one_mean_CI}
\alias{one_mean_CI}
\title{Bootstrap confidence interval for a single mean}
\usage{
one_mean_CI(
  data,
  summary_measure = "mean",
  confidence_level = 0.95,
  number_repetitions = 100
)
}
\arguments{
\item{data}{Vector of quantitative data values.}

\item{summary_measure}{Name of summary measure to return from simulations.
Allowed values are \code{"mean"} or \code{"median"}.
Defaults to \code{"mean"}.}

\item{confidence_level}{Confidence level for interval in decimal form.
Defaults to \code{0.95} (95\% confidence interval).}

\item{number_repetitions}{Number of bootstrapped resamples.}
}
\value{
Returns plot of distribution of bootstrapped statistics,
with values as or more extreme than percentile confidence interval range
highlighted, and reports confidence interval as subtitle on plot.
}
\description{
This function will create a bootstrap confidence interval for a
single mean or median of a quantitative variable.
}
\examples{
set.seed(117)
x <- rnorm(25)
m <- mean(x)
one_mean_CI(x,
            summary_measure = "mean",
            number_repetitions = 100,
            confidence_level = 0.95
)
}
