% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s216-applets.R
\name{one_mean_test}
\alias{one_mean_test}
\title{Simulation-based hypothesis test for a single mean}
\usage{
one_mean_test(
  data,
  null_value = 0,
  summary_measure = "mean",
  shift = 0,
  as_extreme_as,
  direction = c("greater", "less", "two-sided"),
  number_repetitions = 1,
  add_normal = FALSE
)
}
\arguments{
\item{data}{Vector of quantitative data values.}

\item{summary_measure}{Name of summary measure to return from simulations.
Allowed values are \code{"mean"} or \code{"median"}.
Defaults to \code{"mean"}.}

\item{shift}{Amount to shift data values for bootstrapping of null distribution.}

\item{as_extreme_as}{Value of observed mean.}

\item{direction}{Direction of alternative hypothesis.
Allowed values are \code{"greater"}, \code{"less"}, or \code{"two-sided"}.}

\item{number_repetitions}{Number of simulated samples.}

\item{add_normal}{Logical value indicating whether to superimpose a normal
curve on the histogram. Defaults to FALSE.}
}
\value{
Returns plot of distribution of simulated statistics,
with values as or more extreme than specified
value highlighted, and reports
proportion of simulations as or more extreme than specified
as subtitle on plot.
}
\description{
This function will run a simulation-based hypothesis test for a
single mean or median of a quantitative variable.
}
\examples{
set.seed(117)
x <- rnorm(25)
m <- mean(x)
one_mean_test(x,
              summary_measure = "mean",
              shift = -m,
              as_extreme_as = m,
              direction = "two-sided",
              number_repetitions = 100
)
}
