% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s216-applets.R
\name{one_proportion_bootstrap_CI}
\alias{one_proportion_bootstrap_CI}
\title{Bootstrap confidence interval for one proportion}
\usage{
one_proportion_bootstrap_CI(
  sample_size,
  number_successes,
  confidence_level = 0.95,
  number_repetitions = 100
)
}
\arguments{
\item{sample_size}{Number of trials used to compute proportion.}

\item{number_successes}{How many successes were observed in those trials?}

\item{confidence_level}{Confidence level for interval in decimal form.
Defaults to 0.95 (95\% confidence interval).}

\item{number_repetitions}{Number of bootstrapped resamples.}
}
\value{
Returns plot of distribution of bootstrapped statistics,
with values as or more extreme than percentile confidence interval range
highlighted, and reports confidence interval as subtitle on plot.
}
\description{
This function will create a bootstrap confidence interval for a
single proportion of successes.
}
\examples{
one_proportion_bootstrap_CI(
  sample_size = 150,
  number_successes = 98,
  confidence_level = 0.99,
  number_repetitions = 1000
)
}
