% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s216-applets.R
\name{paired_observed_plot}
\alias{paired_observed_plot}
\title{Plots for observed matched pairs data}
\usage{
paired_observed_plot(data, which_first = 1)
}
\arguments{
\item{data}{Data frame
with values for each group in the first two columns.}

\item{which_first}{Which column is first in order of subtraction?
\code{1} if subtracting second column from first (1 - 2);
\code{2} if subtracting first column from second (2 - 1). Defaults to \code{1}.}
}
\value{
Returns plot of distribution of observed values,
with means & standard deviations in groups and paired observations
linked by a line segment, and histogram of differences.
}
\description{
This function will create two plots of a quantitative variable
and its differences for matched pairs data.
}
\examples{
set.seed(117)
x <- rnorm(25)
y <- x + 1 + rnorm(25, 0, 1.8)
data <- data.frame(x, y)
paired_observed_plot(data, which_first = 1)
}
