% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s216-applets.R
\name{regression_bootstrap_CI}
\alias{regression_bootstrap_CI}
\title{Bootstrap confidence interval for regression}
\usage{
regression_bootstrap_CI(
  formula,
  data,
  summary_measure = c("slope", "correlation"),
  confidence_level = 0.95,
  number_repetitions = 100
)
}
\arguments{
\item{formula}{Formula of the form \code{response ~ predictor},
where \code{predictor} defines a quantitative explanatory variable and
\code{response} is a quantitative response variable.}

\item{data}{Data frame with columns for response and predictor variables.}

\item{summary_measure}{Name of summary measure to return from simulations.
Allowed values are
\code{"slope"} for test of slope or
\code{"correlation"} for test of correlation.}

\item{confidence_level}{Confidence level for interval in decimal form.
Defaults to 0.95 (95\% confidence interval).}

\item{number_repetitions}{Number of bootstrapped resamples.}
}
\description{
Function to create bootstrap confidence interval for slope of simple linear regression or correlation between two quantitative variables
}
\examples{
data(mtfires)
mtfires$logHect  <- log(mtfires$Hectares)
regression_bootstrap_CI(logHect ~ Temperature,
  data = mtfires,
  summary_measure = "correlation",
  confidence_level = 0.9,
  number_repetitions = 1000)
}
