% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s216-applets.R
\name{relative_risk_bootstrap_CI}
\alias{relative_risk_bootstrap_CI}
\title{Bootstrap confidence interval for a relative risk}
\usage{
relative_risk_bootstrap_CI(
  formula,
  data,
  group_numerator,
  response_value_numerator,
  confidence_level = 0.95,
  number_repetitions = 100,
  bins = "FD",
  bin_breaks = NULL,
  adjust = FALSE,
  annotate = TRUE,
  show_vlines = TRUE,
  show_rug = TRUE
)
}
\arguments{
\item{formula}{Formula of the form \code{response ~ predictor},
where \code{predictor} defines the two groups of the explanatory variable and
\code{response} is binary or a two-level categorical variable.}

\item{data}{Data frame with columns for response and predictor variables.}

\item{group_numerator}{Value of predictor variable
that should be in the numerator of the relative risk}

\item{response_value_numerator}{Value of response that corresponds
to "success" when computing proportions.}

\item{confidence_level}{Confidence level for interval in decimal form.
Defaults to 0.95 (95\% confidence interval).}

\item{number_repetitions}{Number of bootstrapped resamples.}

\item{bins}{breaks information for hist, defaults to "FD"}

\item{bin_breaks}{vector of break points for bins, defaults to NULL}

\item{adjust}{Adds 0.5 to all counts in the contingency table to avoid RR of infinity
(modifies both observed and bootstrap RRs), defaults to FALSE}

\item{annotate}{add numerical information to plot, defaults to TRUE}

\item{show_vlines}{add CI lines and line for RR = 1 to plot in red, defaults to TRUE}

\item{show_rug}{add alpha tick marks along x-axis for bootstrap RR values,
defaults to TRUE}
}
\value{
Returns plot of distribution of bootstrapped statistics,
with values as or more extreme than percentile confidence interval range
highlighted, and reports confidence interval as subtitle on plot.
}
\description{
This function will create a bootstrap confidence interval for
the ratio of two proportions of successes between two independent groups.
}
\details{
Note that infinite RR values are not displayed but are included in the CI
and noted in annotation. The adjust option avoids infinite RR values by
adding 0.5 to all counts in the 2x2 contingency table.

Some code for this function was modified using perplexity.ai.
}
\examples{
library(ggplot2)
library(dplyr)
exp_gp  <- rep(c("vaccine", "placebo"), c(14, 6))
outcome <- c(rep(c('infection', 'no infection'), c(5, 9)),
             rep(c('infection', 'no infection'), c(6, 0)))
malaria <- data.frame(exp_gp, outcome)
relative_risk_bootstrap_CI(outcome ~ exp_gp,
                                 data = malaria,
                                 group_numerator = "vaccine",
                                 response_value_numerator = "infection",
                                 number_repetitions = 100)
relative_risk_bootstrap_CI(outcome ~ exp_gp,
                                 data = malaria,
                                 group_numerator = "placebo",
                                 response_value_numerator = "infection",
                                 number_repetitions = 500)

}
