% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s216-applets.R
\name{relative_risk_test}
\alias{relative_risk_test}
\title{Simulation-based hypothesis test for a relative risk}
\usage{
relative_risk_test(
  formula,
  data,
  group_numerator,
  response_value_numerator,
  number_repetitions = 1000,
  bins = "FD",
  bin_breaks = NULL,
  as_extreme_as = NULL,
  shade = T,
  direction = c("greater", "less", "two-sided"),
  adjust = FALSE,
  annotate = TRUE,
  show_vlines = TRUE,
  show_rug = TRUE,
  seed = NULL
)
}
\arguments{
\item{formula}{Formula of the form \code{response ~ predictor},
where \code{predictor} defines the two groups of the explanatory variable and
\code{response} is binary or a two-level categorical variable.}

\item{data}{Data frame with columns for response and predictor variables.}

\item{group_numerator}{Value of predictor variable
that should be in the numerator of the relative risk}

\item{response_value_numerator}{Value of response that corresponds
to "success" when computing proportions.}

\item{number_repetitions}{Number of simulated samples.}

\item{bins}{breaks information for hist, defaults to "FD"}

\item{bin_breaks}{vector of break points for bins, defaults to NULL}

\item{as_extreme_as}{Value of observed relative risk,
defaults to NULL and then is calculated using provided data.}

\item{shade}{Logical value indicating whether to shade p-value area.
Defaults to TRUE.}

\item{direction}{Direction of alternative hypothesis.
Allowed values are \code{"greater"}, \code{"less"}, or \code{"two-sided"}.}

\item{adjust}{Adds 0.5 to all counts in the contingency table to avoid RR of infinity
(modifies both observed and permuted RRs), defaults to FALSE}

\item{annotate}{add numerical information to plot, defaults to TRUE}

\item{show_vlines}{add observed RR line and line for RR = 1 to plot in red, defaults to TRUE}

\item{show_rug}{add alpha tick marks along x-axis for bootstrap RR values,
defaults to TRUE}

\item{seed}{Set random number seed for stochastic components for reproducibility,
defaults to NULL and so does not set a seed.}
}
\value{
Returns plot of distribution of simulated statistics,
with values as or more extreme than specified
value highlighted, and reports
proportion of simulations as or more extreme than specified
as subtitle on plot.
}
\description{
This function will run a simulation-based hypothesis test for a
relative risk between two independent groups. It notes any infinite RRs
encountered and allows for an adjustment to avoid this issue.
Infinite RRs are counted in the p-value calculation even though they are
not displayed.
}
\details{
Two-sided p-value is based on doubling the area for direction of
observed RR versus 1.

Some code for this function was modified using perplexity.ai.
}
\examples{
library(dplyr)
library(ggplot2)
exp_gp  <- rep(c("vaccine", "placebo"), c(14, 6))
outcome <- c(rep(c('infection', 'no infection'), c(5, 9)),
             rep(c('infection', 'no infection'), c(6, 0)))
 malaria <- data.frame(exp_gp = factor(exp_gp), outcome = factor(outcome))
 relative_risk_test(outcome ~ exp_gp,
      data = malaria,
      group_numerator = "vaccine",
      response_value_numerator = "infection",
      annotate = T,
      number_repetitions = 1000,
      direction = "two-sided", adjust = F)
 relative_risk_test(outcome ~ exp_gp,
      data = malaria,
      group_numerator = "vaccine",
      response_value_numerator = "infection",
      annotate = T,
      number_repetitions = 1000,
      direction = "less",
      adjust = T)

}
