% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-treadmill.R
\docType{data}
\name{treadmill}
\alias{treadmill}
\title{Fitness Measurements Taken on a Treadmill}
\format{
a \code{data.frame} with 31 observations on 8 variables:
\describe{
\item{Subject}{an integer denoting the person measured}
\item{TreadMillOx}{subjects' treadmill oxygen consumption in ml per kg per
minute}
\item{TreaMillMaxPulse}{subjects' maximum pulse rate in beats per minute}
\item{RunTime}{subjects' minutes to run 1.5 miles}
\item{RunPulse}{subjects' maximum pulse rate during 1.5 mile run in beats per minute}
\item{RestPulse}{subjects' resting pulse rate in beats per minute}
\item{BodyWeight}{subjects' weight in kg}
\item{Age}{subjects' age in years}
}
}
\usage{
treadmill
}
\description{
These data come from a study on (what was this study on?).  Used in Chapter 0 to help introduce
students to R and RStudio.
}
\references{
Westfall, Peter H., S. Stanley Young, and S. Paul Wright.
"On adjusting P-values for multiplicity." Biometrics (1993): 941-945.
}
\keyword{data}
\keyword{fitness}
\keyword{treadmill}
