% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s216-applets.R
\name{two_mean_test}
\alias{two_mean_test}
\title{Simulation-based hypothesis test for a difference in means}
\usage{
two_mean_test(
  formula,
  data,
  first_in_subtraction,
  as_extreme_as,
  direction = c("greater", "less", "two-sided"),
  number_repetitions = 1
)
}
\arguments{
\item{formula}{Formula of the form \code{response ~ predictor},
where \code{predictor} defines the two groups of the explanatory variable and
\code{response} is a quantitative response variable.}

\item{data}{Data frame with columns for response and predictor variables.}

\item{first_in_subtraction}{Value of predictor variable
that should be first in order of subtraction for computing
difference in means.}

\item{as_extreme_as}{Value of observed difference in means.}

\item{direction}{Direction of alternative hypothesis.
Allowed values are \code{"greater"}, \code{"less"}, or \code{"two-sided"}.}

\item{number_repetitions}{Number of simulated samples.}
}
\value{
Produces side-by-side boxplots of observed data and plot of distribution of simulated values, with values as or more extreme than specified value highlighted and count/proportion of those values reported as subtitle on plot
}
\description{
Simulation-based hypothesis test for a difference in means
}
\examples{
data(pt)
two_mean_test(responses ~ brand,
  data = pt,
  first_in_subtraction = "B1",
  as_extreme_as = -.4,
  direction = "two-sided",
  number_repetitions = 1000)
}
