% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s216-applets.R
\name{two_proportion_bootstrap_CI}
\alias{two_proportion_bootstrap_CI}
\title{Function to build bootstrap confidence interval for the difference between two proportions}
\usage{
two_proportion_bootstrap_CI(
  formula,
  data,
  first_in_subtraction,
  response_value_numerator,
  number_repetitions = 1,
  confidence_level = 0.95
)
}
\arguments{
\item{formula}{Formula of the form response~predictor, where predictor defines two groups and response is binary or two-level categorical}

\item{data}{Dataset with columns for response and predictor variable}

\item{first_in_subtraction}{Value of predictor that should be first in order of subtraction for computing statistics}

\item{response_value_numerator}{Value of response that corresponds to "success" computing proportions}

\item{number_repetitions}{number of draws for bootstrap distribution}

\item{confidence_level}{confidence level to use for interval construction in decimal form.  Default is 95\%}
}
\value{
Produces plot of distribution of bootstrapped values, with values as or more extreme than confidence interval range highlighted and reports CI as subtitle on plot
}
\description{
Function to build bootstrap confidence interval for the difference between two proportions
}
\examples{
data(pt)
pt$twoSeconds <- ifelse(pt$responses >=2, "Yes", "No")
two_proportion_bootstrap_CI(twoSeconds~brand, data = pt, first_in_subtraction = "B1", response_value_numerator = "Yes", number_repetitions = 100, confidence_level = 0.95)

}
