% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction-plot-array.R
\name{intplotarray}
\alias{intplotarray}
\title{Interaction Plots array with Error Bars}
\usage{
intplotarray(
  formula = NULL,
  data = NULL,
  type = "b",
  x.cont = FALSE,
  legend = TRUE,
  trace.label = NULL,
  leg.lab = NULL,
  fixed = FALSE,
  x.leg = NULL,
  y.leg = NULL,
  cex.leg = 1,
  ncol = 1,
  pch = c(16, 21, 15, 22, 17, 24, c(3:14)),
  fun = function(x) mean(x, na.rm = TRUE),
  ci.fun = function(x) c(fun(x) - se(x), fun(x) + se(x)),
  err.width = if (length(levels(as.factor(x.factor))) > 10) 0 else 0.1,
  err.col = col,
  err.lty = 1,
  xlim = NULL,
  ylim = NULL,
  cex = NULL,
  lwd = NULL,
  col = 1:10,
  cex.axis = 1,
  xaxt = "s",
  main = NULL,
  cld = F,
  bw = "SJ-dpi",
  cldshift = 0.1,
  cldcol = "white",
  ...
)
}
\arguments{
\item{formula}{an object of class "\link{formula}" (or one that can be
coerced to that class); a symbolic description of the model to be plotted.
The details of the model specification are given under "Details".}

\item{data}{a required data frame containing the variables for the plot.}

\item{type}{character.  One of "p" or "b".  Defaults to "b" for base version that adds lines between points. Choose "p" to not have lines (just points with error bars).}

\item{x.cont}{boolean.  Defaults to FALSE as current interaction plots are not for continuous predictors but included for future development.}

\item{legend}{boolean.  Defaults to TRUE. Adds legend to plot, removal causes other changes.}

\item{trace.label}{text. Adds a text label to the legend}

\item{leg.lab}{text. Changes the names of levels used in legend, must match the number of levels.}

\item{fixed}{boolean. Defaults to FALSE and relates to order of labels in legend.}

\item{x.leg}{Defaults to NULL. Determines x-location of legend on plot.}

\item{y.leg}{Defaults to NULL. Determines x-location of legend on plot.}

\item{cex.leg}{Defaults to 1. Size of text in legend.}

\item{ncol}{Defaults to 1. Number of columns for the legend. 1 is usually best.}

\item{pch}{Symbols for means in plot. Match number of levels as used in first variable in formula.}

\item{fun}{Function for finding the point estimates, defaults to finding the mean. Change not recommended.}

\item{ci.fun}{Function for finding the width of error bars, with default of plus/minus 1 SE. See function for details to change this.}

\item{err.width}{Defaults to 0.1 if less than 10 levels and 0 otherwise. Can specify wider or smaller than 0.1.}

\item{err.col}{Defaults to matching order of levels and other aspects of plot but can make contrasting error bars with other choices.}

\item{err.lty}{Defaults to 1. Line type for the error bars. Best to leave solid but can be changed.}

\item{xlim}{Defaults to NULL. Can make other choices to modify default choice that tries to make room for standard versions of interaction plots.}

\item{ylim}{Defaults to NULL. Can make other choices to modify default choice that tries to make room for standard versions of interaction plots.}

\item{cex}{Defaults to 1. Size of points for means in plot.}

\item{lwd}{Deafults to 1. Line width for lines in plot.}

\item{col}{Defaults to 1:10 for the colors used in the plot.}

\item{cex.axis}{Defaults to 1. Size of text for tick marks on x and y axis.}

\item{xaxt}{Defaults to "s". Change to "n" to remove x-axis tick marks.}

\item{main}{Title for plot.}

\item{cld}{boolean.  Defaults to FALSE. If true, adds compact letter display from Tukey's HSD that is run.}

\item{bw}{Defaults to "SJ-dpi". Option to pass into pirateplots for making the density curves and can be manually specified if there is no variation in a group to get something to display.}

\item{cldshift}{Defaults to 0.1. Amount to shift letters added to plot.}

\item{cldcol}{Colors for letters for CLD.}

\item{...}{optional arguments to be passed to plot.}
}
\description{
This function allows you to generate interaction plots for a Two-Way ANOVA
with error bars with an array of plots and main effect type plots.
}
\details{
Function for making nice looking interaction plots with 1 SE-based intervals, making both versions of interaction plots and adds pirate-plots on the off-diagonal. Can add Tukey's HSD compact letter displays to two interaction plots.
}
\examples{
# must have the carData package installed
data(TitanicSurvival, package = "carData")
intplotarray(age ~ sex * survived, data = TitanicSurvival)
}
