% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s216-applets.R
\name{one_proportion_bootstrap_CI}
\alias{one_proportion_bootstrap_CI}
\title{Function to produce bootstrap confidence interval for one proportion}
\usage{
one_proportion_bootstrap_CI(
  sample_size,
  number_successes,
  number_repetitions = 100,
  confidence_level = 0.95
)
}
\arguments{
\item{sample_size}{number of trials used to compute proportion}

\item{number_repetitions}{number of draws for simulation test}

\item{confidence_level}{confidence level for interval in decimal form.  Defaults to 95\%}

\item{number_success}{how many successes were observed in those trials?}
}
\value{
Produces plot of distribution of bootstrapped values, with values as or more extreme than confidence interval range highlighted and reports CI as subtitle on plot
}
\description{
Function to produce bootstrap confidence interval for one proportion
}
\examples{
one_proportion_bootstrap_CI(sample_size = 150, number_successes = 98, number_repetitions = 100, confidence_level = 0.99)
}
