% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s216-applets.R
\name{one_proportion_test}
\alias{one_proportion_test}
\title{Simulation-based hypothesis test for one proportion}
\usage{
one_proportion_test(
  probability_success = 0.5,
  sample_size = 5,
  summary_measure = c("number", "proportion"),
  as_extreme_as,
  direction = c("greater", "less", "two-sided"),
  number_repetitions = 1,
  add_normal = FALSE
)
}
\arguments{
\item{probability_success}{Value between 0 and 1
representing the null hypothesis value for proportion.}

\item{sample_size}{Number of trials used to compute proportion.}

\item{summary_measure}{Name of summary measure to return from simulations.
Allowed values are
\code{"number"} for number of successes or
\code{"proportion"} for proportion of successes.}

\item{as_extreme_as}{Value of observed statistic.
Between 0 and 1 if \code{summary_measure} is \code{"proportion"};
an integer between 1 and \code{sample_size} if \code{summary_measure} is \code{"number"}.}

\item{direction}{Direction of alternative hypothesis.
Allowed values are \code{"greater"}, \code{"less"}, or \code{"two-sided"}.}

\item{number_repetitions}{Number of simulated samples.}

\item{add_normal}{Logical value indicating whether to superimpose a normal
curve on the histogram. Defaults to FALSE.}
}
\value{
Returns plot of distribution of simulated statistics,
with values as or more extreme than specified
value highlighted, and reports
proportion of simulations as or more extreme than specified
as subtitle on plot.
}
\description{
This function will run a simulation-based hypothesis test for a
single proportion of successes.
}
\examples{
one_proportion_test(
  probability_success = 0.5,
  sample_size = 150,
  summary_measure = "proportion",
  as_extreme_as = 0.65,
  direction = "greater",
  number_repetitions = 1000
)
}
