% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s216-applets.R
\name{one_proportion_test}
\alias{one_proportion_test}
\title{One proportion simulation-based hypothesis test}
\usage{
one_proportion_test(
  probability_success = 0.5,
  sample_size = 5,
  number_repetitions = 1,
  as_extreme_as,
  direction = c("greater", "less", "two-sided"),
  report_value = c("number", "proportion")
)
}
\arguments{
\item{probability_success}{value between 0 and 1, represents null hypothesis value for proportion}

\item{sample_size}{number of trials used to compute proportion}

\item{number_repetitions}{number of draws for simulation test}

\item{as_extreme_as}{observed statistic (between 0 and 1)}

\item{direction}{one of "greater", "less", or "two-sided" to give direction of hypothsis test}

\item{report_value}{value to return from simulations - "number" for number of successes or "proportion" for proportion of successes}
}
\value{
Plot of distribution of simulated values, with values as or more extreme than specified value highlighted and reports proportion of successes in simulations as subtitle on plot
}
\description{
This function will run a simulation-based hypothesis test for the value of a single proportion.
}
\examples{
one_proportion_test(probability_success = 0.5, sample_size = 150, number_repetitions = 100,
          as_extreme_as = 0.65, direction = "greater", return_value = "proportion")
}
