% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s216-applets.R
\name{paired_bootstrap_CI}
\alias{paired_bootstrap_CI}
\title{Function to run bootstrap confidence intervalt for matched pairs data}
\usage{
paired_bootstrap_CI(
  data,
  number_repetitions = 100,
  confidence_level = 0.95,
  which_first = 1
)
}
\arguments{
\item{data}{vector of differences or a two- or three-column data frame, with values for each group in last two columns}

\item{number_repetitions}{number of draws for bootstrap simulation}

\item{confidence_level}{confidence level for interval in decimal form.  Defaults to 95\%}

\item{which_first}{For order of subtraction - which column should be first?}
}
\value{
Produces plot of distribution of bootstrapped values, with values as or more extreme than confidence interval range highlighted and reports CI as subtitle on plot
}
\description{
Function to run bootstrap confidence intervalt for matched pairs data
}
\examples{
set.seed(117)
x <- rnorm(25)
y <- x + 1 + rnorm(25, 0, 1.8)
data <- data.frame(x,y)
paired_bootstrap_CI(data, number_repetitions = 100, confidence_level = 0.9, which_first = 1)
}
