% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s216-applets.R
\name{paired_test}
\alias{paired_test}
\title{Function to run simulation test for matched pairs}
\usage{
paired_test(
  data,
  shift = 0,
  direction = c("greater", "less", "two-sided"),
  as_extreme_as,
  number_repetitions = 3,
  which_first = 1
)
}
\arguments{
\item{data}{vector of differences or a two- or three-column data frame, with values for each group in last two columns}

\item{shift}{amount to shift differences for bootstrapping of null distribution}

\item{direction}{one of "greater", "less", or "two-sided" to give direction of hypothsis test}

\item{as_extreme_as}{observed statistic}

\item{number_repetitions}{number of draws for simulation test}

\item{which_first}{For order of subtraction - which column should be first?}
}
\value{
Produces plot of distribution of simulated values, with values as or more extreme than specified value highlighted and count/proportion of those values reported as subtitle on plot
}
\description{
Function to run simulation test for matched pairs
}
\examples{
set.seed(117)
x <- rnorm(25)
y <- x + 1 + rnorm(25, 0, .3)
data <- data.frame(x,y)
obs_diff <- mean(x - y)
paired_test(data, shift = obs_diff, direction = "two-sided", as_extreme_as = obs_diff,
         number_repetitions = 100, which_first = 1)
}
