% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s216-applets.R
\name{regression_bootstrap_CI}
\alias{regression_bootstrap_CI}
\title{Bootstrap confidence intervals for regression}
\usage{
regression_bootstrap_CI(
  formula,
  data,
  confidence_level = 0.95,
  statistic = c("slope", "correlation"),
  number_repetitions = 3
)
}
\arguments{
\item{formula}{Formula of the form response~predictor, where predictor defines two groups and response is binary or two-level categorical}

\item{data}{Dataset with columns for response and predictor variable}

\item{confidence_level}{confidence level to use for interval in decimal form.  Default is for 95\% CI}

\item{statistic}{"slope" for test of slope or "correlation" for test of correlation}

\item{number_repetitions}{number of draws for bootstrap simulation}
}
\description{
Function to create bootstrap confidence interval for slope of simple linear regression or correlation between two quantitative variables
}
\examples{
data(mtfires)
mtfire$logHect  <- log(Hectares)
regression_bootstrap_CI(logHect~Temperature, data = mtfires, statistic = "correlation",
             confidence_level = 0.9, number_repetitions = 1000)
}
