% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s216-applets.R
\name{regression_test}
\alias{regression_test}
\title{Simulation-based hypothesis test for regression}
\usage{
regression_test(
  formula,
  data,
  summary_measure = c("slope", "correlation"),
  as_extreme_as,
  direction = c("greater", "less", "two-sided"),
  number_repetitions = 1
)
}
\arguments{
\item{formula}{Formula of the form \code{response ~ predictor},
where \code{predictor} defines a quantitative explanatory variable and
\code{response} is a quantitative response variable.}

\item{data}{Data frame with columns for response and predictor variables.}

\item{summary_measure}{Name of summary measure to return from simulations.
Allowed values are
\code{"slope"} for test of slope or
\code{"correlation"} for test of correlation.}

\item{as_extreme_as}{Value of observed slope or correlation.}

\item{direction}{Direction of alternative hypothesis.
Allowed values are \code{"greater"}, \code{"less"}, or \code{"two-sided"}.}

\item{number_repetitions}{Number of simulated samples.}
}
\description{
Function to perform simulation-based test for slope of simple linear regression or correlation between two quantitative variables
}
\examples{
data(mtfires)
mtfires$logHect  <- log(mtfires$Hectares)
regression_test(logHect ~ Temperature,
  data = mtfires,
  summary_measure = "correlation",
  as_extreme_as = 1.388,
  direction = "greater",
  number_repetitions = 1000)
}
