% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s216-applets.R
\name{regression_test}
\alias{regression_test}
\title{Simulation-based test for regression}
\usage{
regression_test(
  formula,
  data,
  statistic = c("slope", "correlation"),
  direction = c("greater", "less", "two-sided"),
  as_extreme_as,
  number_repetitions = 3
)
}
\arguments{
\item{formula}{Formula of the form response~predictor, where predictor defines two groups and response is binary or two-level categorical}

\item{data}{Dataset with columns for response and predictor variable}

\item{statistic}{"slope" for test of slope or "correlation" for test of correlation}

\item{direction}{one of "greater", "less", or "two-sided" to give direction of hypothsis test}

\item{as_extreme_as}{observed statistic}

\item{number_repetitions}{number of draws for simulation test}
}
\description{
Function to perform simulation-based test for slope of simple linear regression or correlation between two quantitative variables
}
\examples{
data(mtfires)
mtfire$logHect  <- log(mtfires$Hectares)
regression_test(logHect~Temperature, data = mtfires,
         direction = "greater", statistic = "correlation",
         as_extreme_as = 1.388, number_repetitions = 1000)
}
