% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s216-applets.R
\name{two_mean_bootstrap_CI}
\alias{two_mean_bootstrap_CI}
\title{Bootstrap confidence interval for a difference in means}
\usage{
two_mean_bootstrap_CI(
  formula,
  data,
  summary_measure = "mean",
  first_in_subtraction,
  confidence_level = 0.95,
  number_repetitions = 100
)
}
\arguments{
\item{formula}{Formula of the form \code{response ~ predictor},
where \code{predictor} defines the two groups of the explanatory variable and
\code{response} is a quantitative response variable.}

\item{data}{Data frame with columns for response and predictor variables.}

\item{summary_measure}{Name of summary measure to return from simulations.
Allowed values are
\code{"mean"} for confidence interval for difference in means
\code{"median"} for confidence interval for difference in medians.
Defaults to \code{"mean"}.}

\item{first_in_subtraction}{Value of predictor variable
that should be first in order of subtraction for computing
difference in means.}

\item{confidence_level}{Confidence level for interval in decimal form.
Defaults to 0.95 (95\% confidence interval).}

\item{number_repetitions}{Number of bootstrapped resamples.}
}
\value{
Returns plot of distribution of bootstrapped statistics,
with values as or more extreme than percentile confidence interval range
highlighted, and reports confidence interval as subtitle on plot.
}
\description{
This function will create a bootstrap confidence interval for a
difference in means or difference in medians of a quantitative variable
for two independent groups.
}
\examples{
data(pt)
two_mean_bootstrap_CI(responses ~ brand,
  data = pt,
  first_in_subtraction = "B1",
  confidence_level = 0.98,
  number_repetitions = 1000
)
}
