% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s216-applets.R
\name{two_mean_test}
\alias{two_mean_test}
\title{Function to perform hypothesis test for equality of two means using simulation}
\usage{
two_mean_test(
  formula,
  data,
  first_in_subtraction,
  direction = c("greater", "less", "two-sided"),
  as_extreme_as,
  number_repetitions = 3
)
}
\arguments{
\item{formula}{Formula of the form response~predictor, where predictor defines two groups and response is binary or two-level categorical}

\item{data}{Dataset with columns for response and predictor variable}

\item{first_in_subtraction}{Value of predictor that should be first in order of subtraction for computing statistics}

\item{direction}{one of "greater", "less", or "two-sided" to give direction of hypothsis test}

\item{as_extreme_as}{observed statistic}

\item{number_repetitions}{number of draws for simulation test}
}
\value{
Produces side-by-side boxplots of observed data and plot of distribution of simulated values, with values as or more extreme than specified value highlighted and count/proportion of those values reported as subtitle on plot
}
\description{
Function to perform hypothesis test for equality of two means using simulation
}
\examples{
data(pt)
two_mean_test(responses~brand, data = pt, first_in_subtraction = "B1",
          number_repetitions = 100, as_extreme_as = -.4, direction = "two-sided")

}
